requirements.txt
duckdb>=0.10.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.33.0
plotly>=5.20.0
scipy>=1.10.0
python-dotenv>=1.0.0
openai>=1.30.0
anthropic>=0.25.0
chromadb>=0.4.24
